split dynamic quantize builder family
