+++ onnx2tf/tflite_builder/op_builders/__init__.py
@@ -173,7 +173,6 @@ from onnx2tf.tflite_builder.op_builders.control import (
 from onnx2tf.tflite_builder.op_builders.quantized import (
     build_conv_integer_op,
     build_dequantize_linear_op,
-    build_dynamic_quantize_linear_op,
     build_qgemm_op,
     build_qlinear_add_op,
     build_qlinear_average_pool_op,
@@ -186,6 +185,9 @@ from onnx2tf.tflite_builder.op_builders.quantized import (
     build_qlinear_sigmoid_op,
     build_qlinear_softmax_op,
     build_quantize_linear_op,
+)
+from onnx2tf.tflite_builder.op_builders.dynamic_quantize import (
+    build_dynamic_quantize_linear_op,
 )
 
 __all__ = [
